+++ CHANGELOG.md
@@ -5,6 +5,12 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [5.0.0](../../tags/v5.0.0) - 2024-01-16
+### Changed
+- Update project configurations
+- Update dependencies
+- Update whitelisted hosts
+
 ## [4.0.1](../../tags/v4.0.1) - 2023-11-12
 ### Changed
 - Update dependencies